+++ src/assets/icons/index-svg.ts
@@ -18,6 +18,8 @@ export { ReactComponent as MinusSquare } from "./MinusSquare.svg";
 export { ReactComponent as Paste } from "./Paste.svg";
 export { ReactComponent as Pencil } from "./Pencil.svg";
 export { ReactComponent as PlusSquare } from "./PlusSquare.svg";
+export { ReactComponent as RectFill } from "./RectFill.svg";
+export { ReactComponent as RectFrame } from "./RectFrame.svg";
 export { ReactComponent as Redo } from "./Redo.svg";
 export { ReactComponent as Rename } from "./Rename.svg";
 export { ReactComponent as Trash } from "./Trash.svg";